+++ tests/admin/models/LangTest.php
@@ -1,0 +1,38 @@
+<?php
+
+namespace luya\admin\tests\admin\models;
+
+use admintests\AdminModelTestCase;
+use luya\testsuite\fixtures\NgRestModelFixture;
+use luya\admin\models\Lang;
+use luya\admin\models\NgrestLog;
+
+class LangTest extends AdminModelTestCase
+{
+    public function testEvents()
+    {
+        $fixture = new NgRestModelFixture([
+            'modelClass' => Lang::class,
+        ]);
+
+        $log = new NgRestModelFixture([
+            'modelClass' => NgrestLog::class,
+        ]);
+
+        /** @var Lang $model */
+        $model = $fixture->newModel;
+        $model->short_code = 'fr';
+        $model->name = 'Francais';
+        $model->is_default = 1;
+        $model->is_deleted = 0;
+
+        $this->assertTrue($model->validate());
+        $this->assertTrue($model->insert());
+
+        $model->short_code = 'en';
+        $model->is_default = 0;
+        $this->assertSame(1, $model->update());
+
+        $this->assertTrue($model->delete());
+    }
+}